# swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Map API - GIS
  description: API for retrieving district, municipality and parish GIS information
  version: 1.0.0
servers:
  - url: https://mapas-api.impedro.com
    description: Remote development server
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /api/codigo-postal/{code}:
    get:
      summary: Get zip code information
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{4}-\d{3}$'
      responses:
        '200':
          description: Zip code information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZipCode'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Zip code not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/codigo-postal/{code}/proximidade:
    get:
      summary: Get nearby zip codes by zip code
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{4}-\d{3}$'
      responses:
        '200':
          description: Nearby zip codes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZipCodeList'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No zip codes found nearby
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/codigo-postal/proximidade:
    get:
      summary: Get nearby zip codes by coordinates
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: lng
          in: query
          required: true
          schema:
            type: number
            format: double
      responses:
        '200':
          description: Nearby zip codes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZipCodeList'
        '400':
          description: Missing or invalid latitude/longitude
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No zip codes found nearby
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      example:
        error: "Parameter is missing or invalid"
...
    ZipCode:
      type: object
      properties:
        dtmnfr:
          type: string
        distrito:
          type: string
        municipio:
          type: string
        num_cod_postal:
          type: string
        ext_cod_postal:
          type: string
        cod_postal:
          type: string
        morada:
          type: string
        local_arteria:
          type: string
        porta_inicial:
          type: string
      required: [dtmnfr, cod_postal]
      example:
        dtmnfr: "0101"
        distrito: "Example District"
        municipio: "Example Municipality"
        num_cod_postal: "1234"
        ext_cod_postal: "567"
        cod_postal: "1234-567 Example"
        morada: "Rua Exemplo 1"
        local_arteria: "Bairro"
        porta_inicial: "1A"
    ZipCodeList:
      type: array
      items:
        $ref: '#/components/schemas/ZipCode'
...